- add a new base class for the list widgets ListWidgetBase
