+++ phoenicis-javafx/src/main/java/org/phoenicis/javafx/components/common/control/CombinedListWidget.java
@@ -2,29 +2,18 @@ package org.phoenicis.javafx.components.common.control;
 
 import javafx.beans.property.ObjectProperty;
 import javafx.beans.property.SimpleObjectProperty;
-import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import org.phoenicis.javafx.components.common.skin.CombinedListWidgetSkin;
-import org.phoenicis.javafx.views.common.widgets.lists.ListWidget;
 import org.phoenicis.javafx.views.common.widgets.lists.ListWidgetEntry;
 import org.phoenicis.javafx.views.common.widgets.lists.ListWidgetType;
 
-public class CombinedListWidget<E> extends ControlBase<CombinedListWidget<E>, CombinedListWidgetSkin<E>> {
-    /**
-     * A list containing all elements to be shown inside this {@link ListWidget}
-     */
-    private final ObservableList<ListWidgetEntry<E>> elements;
-
-    private final ObjectProperty<ListWidgetEntry<E>> selectedElement;
-
+public class CombinedListWidget<E> extends ListWidgetBase<E, CombinedListWidget<E>, CombinedListWidgetSkin<E>> {
     private final ObjectProperty<ListWidgetType> selectedListWidget;
 
     public CombinedListWidget(ObservableList<ListWidgetEntry<E>> elements,
             ObjectProperty<ListWidgetEntry<E>> selectedElement, ObjectProperty<ListWidgetType> selectedListWidget) {
-        super();
+        super(elements, selectedElement);
 
-        this.elements = elements;
-        this.selectedElement = selectedElement;
         this.selectedListWidget = selectedListWidget;
     }
 
@@ -32,29 +21,9 @@ public class CombinedListWidget<E> extends ControlBase<CombinedListWidget<E>, Co
         this(elements, new SimpleObjectProperty<>(), new SimpleObjectProperty<>());
     }
 
-    public CombinedListWidget() {
-        this(FXCollections.observableArrayList(), new SimpleObjectProperty<>(), new SimpleObjectProperty<>());
-    }
-
     @Override
     public CombinedListWidgetSkin<E> createSkin() {
         return new CombinedListWidgetSkin<>(this);
-    }
-
-    public ObservableList<ListWidgetEntry<E>> getElements() {
-        return elements;
-    }
-
-    public ListWidgetEntry<E> getSelectedElement() {
-        return selectedElement.get();
-    }
-
-    public ObjectProperty<ListWidgetEntry<E>> selectedElementProperty() {
-        return selectedElement;
-    }
-
-    public void setSelectedElement(ListWidgetEntry<E> selectedElement) {
-        this.selectedElement.set(selectedElement);
     }
 
     public ListWidgetType getSelectedListWidget() {
