+++ phoenicis-javafx/src/main/java/org/phoenicis/javafx/components/common/control/IconsListWidget.java
@@ -3,43 +3,16 @@ package org.phoenicis.javafx.components.common.control;
 import javafx.beans.property.ObjectProperty;
 import javafx.collections.ObservableList;
 import org.phoenicis.javafx.components.common.skin.IconsListWidgetSkin;
-import org.phoenicis.javafx.views.common.widgets.lists.ListWidget;
 import org.phoenicis.javafx.views.common.widgets.lists.ListWidgetEntry;
 
-public class IconsListWidget<E> extends ControlBase<IconsListWidget<E>, IconsListWidgetSkin<E>> {
-    /**
-     * A list containing all elements to be shown inside this {@link ListWidget}
-     */
-    private final ObservableList<ListWidgetEntry<E>> elements;
-
-    private final ObjectProperty<ListWidgetEntry<E>> selectedElement;
-
+public class IconsListWidget<E> extends ListWidgetBase<E, IconsListWidget<E>, IconsListWidgetSkin<E>> {
     public IconsListWidget(ObservableList<ListWidgetEntry<E>> elements,
             ObjectProperty<ListWidgetEntry<E>> selectedElement) {
-        super();
-
-        this.elements = elements;
-        this.selectedElement = selectedElement;
+        super(elements, selectedElement);
     }
 
     @Override
     public IconsListWidgetSkin<E> createSkin() {
         return new IconsListWidgetSkin<>(this);
-    }
-
-    public ObservableList<ListWidgetEntry<E>> getElements() {
-        return elements;
-    }
-
-    public ListWidgetEntry<E> getSelectedElement() {
-        return selectedElement.get();
-    }
-
-    public ObjectProperty<ListWidgetEntry<E>> selectedElementProperty() {
-        return selectedElement;
-    }
-
-    public void setSelectedElement(ListWidgetEntry<E> selectedElement) {
-        this.selectedElement.set(selectedElement);
     }
 }
